requirements.txt
fastapi
uvicorn[standard]
numpy
Pillow
deepface
tensorflow[and-cuda]
torch
mtcnn
tf-keras